import: Move 'process_avatars' to import_util.
