+++ tools/test-backend
@@ -36,7 +36,6 @@ target_fully_covered = {path for target in [
     'zproject/backends.py',
     'confirmation/*.py',
     'zerver/webhooks/*/*.py',
-    'zerver/data_import/import_util.py',
     # Once we have a nice negative tests system, we can add these:
     # 'zerver/webhooks/*/*.py',
     # 'zerver/webhooks/*/*/*.py',
@@ -99,6 +98,7 @@ not_yet_fully_covered = {
     # Data import files
     'zerver/data_import/slack.py',
     'zerver/data_import/gitter.py',
+    'zerver/data_import/import_util.py',
     # Webhook integrations with incomplete coverage
     'zerver/webhooks/beanstalk/view.py',
     'zerver/webhooks/bitbucket2/view.py',
